[Refactor] Move JDBC type checker out of Types module
