+++ be/src/exec/jdbc/type_checker.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "types/checker/type_checker.h"
+#include "exec/jdbc/type_checker.h"
 
 #include <fmt/format.h>
