+++ be/src/exec/jdbc/type_checker.h
